NPC Collision Update
Refactored NPC code, cut it down a bit.
NPC now stops and looks at the player properly when the player approaches.
NPC has option to toggle if collisions are disabled from the start
NPC has option to disable collisions after a the player has been detected a certain amount of time so they don't get stuck in the field.