Found and fixed memory leak issue
Cause of memory leak was found to be within enemy state scripts that used the code await get_tree().create_timer().timeout
The reason for the memory leak was found to be that the timers were being created outside of the scene tree, therefore would not be deleted when queue_free() was called on the enemy node. Since idle or walk states were the only states able to be used by the enemies at the this time, there was always a timer of some sort going.

New code implemented to fix the memory leak. In short, created a new timer with code but the timer is now added as a child of the enemy node so it properly destroys on queue_free()

These changes were also made to the NPC, since they could potentially have the same issues due to their state code being almost identical.