Dialogue Manager Update 2.0
Updated dialogue system to allow characters to talk with sound
Update dialogue system to allow portraits to animate
Created Options global so portrait animation and voices to be full, semi, or off.
Added a bunch of placeholder sounds for testing. (to be removed later)

Updated DialogueManager scripts with #region markers to help determine what code was added to the original DialogueManager scripts (balloon, dialogu_manager)