Item Use Update
top_level_stats updated to allow party members to be selected within the menu
top_level_stats updated to reflect the vitals of the party members
-- status effects remain to be implemented --
item_effect updated to allow them to be used through the menu (but will need further enhancements later for battle, likely)
player_path.gd / tscn deleted since it is not being used in any way to determine pathing
player_character updated to disallow movement while the menu is open
all party member's data tweaked for testing
options sliders no longer handle input, may need to revisit the way input is handled later. workaround was to simply await a process frame (a bug existed that made it so the options menu was completely exited when hitting cancel_input after selecting a slider)
inventory_item_button updated to also hold a reference to the item slot it came from in Inventory.current_inventory (in order to decrement item quantities in that slot)
inventory_button given more focus functionality so the menu was not as confusing

game_menu_gui:
added variables to keep track of various "last button" needs in order to refocus them after the menu was closed
Filled out documentation
added ability to pull information from CharDataKeeper.party_members to instantiate top_level_stats scenes
added tons of functions to handle using these scenes (selection, passing information through, etc)
added functionality for when an item is completely used up, another item is focused (instead of focusing the inventory options menu portion again, which is annoying)
returned functionality to  update_item_description() to properly hide the equip portion of the description if the item was not a piece of equipment, and also returned the highlighting functionality.
Added an AudioStreamPlayer to the scene to allow sounds to be played later

Updated some code (player and game menu) to start utilizing game_state. determined keeping this as an int was the best way to go.