Main Created, Title Created, Scene Manager Updated
Main scene created in order to control the way battle and field are shown. (this can be done with autoloads but it is probably easiest to do with a main scene controlling things instead
barebones title scene created in order to be instantiated when the game opens into Main.
refactored some code in scene_manager and npc to take care of the swap to a Main scene setup.