SceneManager, SceneTransitioner Reconfigure
Reconfigured the code within scenemanager to be easier to use
(some work needs to be done to instantiate scenes that do not use a SceneTransitioner, EX: a cutscene)
Reconfigured the SceneTransitioner code to work with this (and much simpler), most things are handled by SceneManager now
Updated PlayerCharacter to force idle state when game menu opens
Updated PlayerCharacter to always run when Options.always_run is true, and the opposite
player_walk updated for Options.always_run functionality. if Options.always_run is true, the cancel button will cause the player to walk
player_idle updated to have a function to force_idle (called when game menu opened)

Dev Menu button 1 now instantiates the test scene Test01 and instantiates the player from CharDataKeeper.party_members[0].

FieldCameraRig within Main given a script to allow it to follow the player. Previously, the active camera was just part of the player scene (which was too static)

TemplateScene updated to have a few more nodes just in case.
TemplateScene Y Sorting enabled.

transition_layer's gradient altered so all colors are black, and the two edges are just more transparent black. The white looked bad and was very distracting, black causes it to look more like a proper fade.