Enemy / Enemy Spawner redo
Decided the best way to handle enemies was to configure them as individual scenes instead of trying to pull data from a resource and use a base enemy scene.
Refactored EnemySpawner to work with this
Created EnemyGroup class type to handle this
Deleted BattleEnemy script due to the enemies now being held within their own scenes, which will hold both field and battle sprites/animations
FieldScene had some functions that were not needed, commented out for the time being.
LootDrop class added in order to handle which item and the chance it is dropped (or stolen). Function created in loot drop to help determine if this happens.
CharResource updated to take out the enemy variables. Other variables will be taken out and moved into their respective scenes so things can be more individually tuned.